fixed several bugs in the DexReader that was preventing some of the APAC APK files from being read.  these changes have now been tested on a much larger set of APK, which also includes those that we have from a malware posting.    src/main/scala/com/ucombinator/dalvik/AST/DexReader.scala,    src/main/scala/com/ucombinator/dalvik/Analyzer.scala,    src/main/scala/com/ucombinator/dalvik/android/ApkReader.scala
lowercased the initializer variables of the instruction case classes to make it
easier to cut and paste them into pattern matches.  (the uppercased variables
apparently foul up the scala pattern matcher, I suspect because they are
interpreted as types by the parser.)
   src/main/scala/com/ucombinator/dalvik/AST/Instruction.scala

added the first (very, very) simple analysis.  basically, you can give the
analyzer a class name and method name and it gives you a list of all possible
methods called from the procedure.  (it shouldbe noted that this list is just
the list of archetype methods, since a child class could override the method.)
   src/main/scala/com/ucombinator/dalvik/analysis/TrackMethodCalls.scala (new)
   src/main/scala/com/ucombinator/dalvik/Analyzer.scala,