Added an initial (very, very) simple analyzer and fixed a couple bugs.
fixed bug when returning full methods and fields lists from a ClassDef when one
of the instanceFields, staticFields, directMethods, and virtualMethods is null.
   src/main/scala/com/ucombinator/dalvik/AST/ClassDef.scala

added method to lookup the class associated with a method.
   src/main/scala/com/ucombinator/dalvik/AST/MethodDef.scala

created an analyzer to identify which methods contain sources and sinks based
on the admittedly limited and possibly not quite correct list we currently have
on hand.
   src/main/scala/com/ucombinator/dalvik/Analyzer.scala,
   src/main/scala/com/ucombinator/dalvik/analysis/TrackMethodCalls.scala