added line numbers to results; made source/sinks specifiable; updated code to use scala 2.10.1
changed code to use Scala 2.10.1 and fixed deprecation/feature warnings
   build.sbt, DexReader.scala

added source location information to results.  looks up the nearest
source location in the debug information for the instructions in the
method.
   MethodDef.scala, Analyzer.scala

made sources, sinks, and other methods we might be interested in
specifiable in an xml file.  the end user can also specify a different
xml file to use to control the lists of sources, sinks, and other
interesting methods.  the current config file is based on what is
posted on the DARPA page with this information.
   Analyzer.scala, TrackMethodCalls.scala, config/sourceSink.xml (new)

also added a shell script wrapping the analyzer called analyzer in the
main directory.
   analyzer (new)