new class to build simple method call graph
the new method call graph is built by determining where a method is
called and at each call site, the set of methods that could potentially
be called there.  right now the code does not take into account the
reflection API's ability to call arbitrary methods (or the ClassLoaders
ability to load new functions).  when these things happen we lose
soundness of the analysis.  (the former is easier to catch, but the
latter might not be).
   SimpleMethodCallGraph.scala (new)