updated existing features to use the new SimpleMethodCallGraph
updated the existing source, sink, and other interesting method
reporting to use the SimpleMethodCallGraph as its starting point.  the
reporting of the methods called from a given method (when a class name
and method name are given) is now also driven on this.  there is a lot
more information we could query from this model and it is exceedingly
fast to build it, so it is a good start at the next step.
    Analyzer.scala, TrackMethodCalls.scala